cache: cache-position needs to be int64_t
Both max_filepos and offset are int64_t,
so pos can overflow, e.g. causing endless loops
in stream implementation.

+++ stream/cache.c
@@ -241,7 +241,7 @@ static bool cache_fill(struct priv *s)
     int64_t space = s->buffer_size - (newb + back);
 
     // offset into the buffer that maps to max_filepos
-    int pos = s->max_filepos - s->offset;
+    int64_t pos = s->max_filepos - s->offset;
     if (pos >= s->buffer_size)
         pos -= s->buffer_size; // wrap-around
